🐛 Fix case where no solution is found within time limit in first batch
Example case which is fixed by this commit:

If Baron is the only solver which is requested for solving and time limit is set to value T.
Now, if no solution is found by Baron with any model within T, then Baron would exit automatically because of time limit set by the options.

We want execution of first batch to continue until a solution is found by any one instance of the solver.