🐛 Fix the case where the program enters into an infinite loop
The infinite loop had occurred when Octeract was used as the solver and internet connection was not available. The solver did not immediately exit, instead it tried to connect to its server couple of times and once its threshold was reached, it exited without solving the graph/network (i.e. data/input testcase). And, while the solvers and their corresponding `tmux` sessions ended, the program was thinking that they are executing but could not find a feasible solution and hence it would keep giving them extra time.

The `0_status` file message corresponding to this situation has also been updated.